fixing helper function, organize

+++ app/rental_routes.py
@@ -10,24 +10,24 @@ rentals_bp = Blueprint("rentals",__name__, url_prefix="/rentals")
 def handle_rental(register):
     if register == "check-in":
         request_body = request.get_json()
-        validates_rental_request_body(request_body)
-        cust_id = request_body["customer_id"]
-        vid_id = request_body["video_id"]
+        cust_id, vid_id = validates_request_body(request_body)
+
         customer = Customer.query.get(cust_id)
         video = Video.query.get(vid_id)
-        rental = Rental.query.filter_by(video_id=vid_id,customer_id=cust_id)
+
+        rental = db.session.query(Rental).filter(Rental.customer_id == cust_id, Rental.video_id == vid_id).first()
+        if not rental:
+            return make_response({"message": "No outstanding rentals for customer 1 and video 1"}, 400)
+
         db.session.delete(rental)
+        customer.videos_checked_out_count -= 1
         db.session.commit()
         return make_response(jsonify(create_checkout_dict(video,customer)),200)
 
-
-
     elif register == "check-out":
         request_body = request.get_json()
-        validates_rental_request_body(request_body)
-        video_id = request_body["video_id"]
+        customer_id, video_id = validates_request_body(request_body)
         video = Video.query.get(video_id)
-        customer_id = request_body["customer_id"]
         customer = Customer.query.get(customer_id)
 
         rental = Rental(
@@ -52,7 +52,7 @@ def create_checkout_dict(video, customer):
     "available_inventory" : video.gets_available_inventory()}
     return result
 
-def validates_rental_request_body(request_body):
+def validates_request_body(request_body):
     if "customer_id" not in request_body or "video_id" not in request_body:
         abort(make_response(jsonify({"error" : "Missing customer_id or video_id"}),400))
     customer_id = request_body["customer_id"]
@@ -63,5 +63,7 @@ def validates_rental_request_body(request_body):
     video_id = request_body["video_id"]
     video = Video.query.get(video_id)
     if not video:
-        abort(make_response("Video not found", 404))  
+        abort(make_response("Video not found", 404))
+
+    return customer_id, video_id
